fix(node): 节点扩缩容逻辑优化
