+++ locales/en-US/home.yml
@@ -139,3 +139,5 @@ USER_HAS_NO_RESOURCE: "user has no resource"
 
 #storage
 PROVISIONER_EXSIT: "Storage provider already exists"
+
+TASK_IN_EXECUTION: "A task is being executed in the current cluster. Please try again later..."
